pyproject.toml
[tool.poetry]
name = "pxls-auth"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.20.0"
sanic = "^22.3.2"
aiohttp = "^3.8.1"
lxml = "^4.9.0"
expiring-dict = "^1.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.3.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length=100